Make LwM2mModel an interface. Concrete class is now StaticModel
